requirements.txt
langchain
langchain-community
langchain-groq
faiss-cpu
pypdf
python-dotenv
sentence-transformers
langchain-text-splitters
fastapi uvicorn
